WIP use StackDependencyGraph to handle dependencies
A work in progress commit that does not yet fully work when using
StackDependencyGraph to launch and delete stacks. The correct DAG is
produced but since we are storing the path rather than Stack object the
@recurse_into_sub_stack_group decorator does not run fully. in order to
get this to work we need to get the Stack object into the graph.